Criando a logica do exerc2
