+++ semana24/Aula64/src/exercicio2/index.ts
@@ -1,0 +1,10 @@
+import { handleString } from "../functions/HandleString"
+
+
+const compressString = (text: string): string => {
+
+    return text
+}
+
+compressString("Tese")
+
