+++ semana24/Aula64/src/functions/HandleString.ts
@@ -1,0 +1,5 @@
+export const handleString = (text: string) => {
+    return text
+        .normalize('NFD')
+        .replace(/([\u0300-\u036f]|[^0-9a-zA-Z])/g, "")
+}
